requirements.txt
fastapi
uvicorn
pydantic==2.10.6
numpy==1.23.5
pandas==1.5.3
opencv-python-headless==4.11.0.86
Pillow
torch==2.6.0
torchvision==0.21.0
ultralytics==8.3.160
boto3

